wip: project model / auth

+++ backend/nf_cloud_backend/controllers/workflow.py
@@ -21,11 +21,11 @@ from typing import Annotated
 from fastapi import APIRouter, Body, Depends, HTTPException, Header, Response, status
 from pydantic import BaseModel, Field
 from sqlalchemy import select
-from sqlmodel import SQLModel
+from sqlmodel import SQLModel, Session
 
-from .depends import ExistingUser, ExistingWorkflow
+from .depends import Authenticated, AuthenticatedUser, ExistingUser, ExistingWorkflow
 
-from ..models.user import User
+from ..models.user import User, UserRole
 from ..models.workflow import Workflow
 from ..models.workflow_share import WorkflowShare
 
@@ -37,14 +37,35 @@ router = APIRouter(
 )
 
 
+def get_workflow_share(user: User, workflow: Workflow, session: Session) -> WorkflowShare | None:
+    return session.exec(
+        select(WorkflowShare).where(WorkflowShare.user_id == user.id, WorkflowShare.workflow_id == workflow.id)
+    ).one_or_none()[0]
+
+
+def can_access_workflow(user: User, workflow: Workflow, session: Session, for_writing: bool = False) -> bool:
+    match user.role:
+        case UserRole.admin:
+            return True
+        case UserRole.default:
+            if workflow.owner_id == user.id:
+                return True
+            share = get_workflow_share(user, workflow, session)
+            return share is not None and (not for_writing or share.write)
+
+
 @router.get("/",
             summary="List Workflows")
-async def list(session: DbSession) -> list[int]:
+async def list(session: DbSession, auth: Authenticated) -> list[int]:
     """
     Lists the IDs of all workflows visible to this user.
     """
-
-    return [i[0] for i in session.exec(select(Workflow.id)).all()]
+    
+    return [
+        i[0]
+        for i in session.exec(select(Workflow)).all()
+        if can_access_workflow(auth, i[0], session)
+    ]
 
 
 class WorkflowCreateParams(BaseModel):
@@ -55,15 +76,14 @@ class WorkflowCreateParams(BaseModel):
 
 @router.post("/new",
              summary="Create Workflow")
-async def new(params: WorkflowCreateParams, session: DbSession, response: Response) -> int:
+async def new(params: WorkflowCreateParams, session: DbSession, auth: Authenticated) -> int:
     """
     Creates a new workflow with the provided parameters.
-
-    For non-admins, the owner id has to be the currently logged in user.
     """
 
     workflow = Workflow(
         name=params.name,
+        owner_id=auth.id,
         description=params.description,
         definition=params.definition,
         is_published=params.is_published
@@ -76,17 +96,21 @@ async def new(params: WorkflowCreateParams, session: DbSession, response: Respon
 
 class WorkflowShown(BaseModel):
     name: str
+    owner: User | None
     description: str
     definition: dict
     is_published: bool
 
 @router.get("/{workflow_id}",
             summary="Show Single Workflow")
-async def show(workflow: ExistingWorkflow) -> WorkflowShown:
+async def show(workflow: ExistingWorkflow, auth: Authenticated, session: DbSession) -> WorkflowShown:
     """
     Displays information for a single workflow.
     """
 
+    if not can_access_workflow(auth, workflow, session):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+    
     return WorkflowShown(
         name=workflow.name,
         description=workflow.description,
@@ -103,10 +127,13 @@ class WorkflowUpdateParams(BaseModel):
 
 @router.post("/{workflow_id}/edit",
              summary="Edit Workflow")
-async def edit(params: WorkflowUpdateParams, workflow: ExistingWorkflow) -> None:
+async def edit(params: WorkflowUpdateParams, workflow: ExistingWorkflow, auth: Authenticated, session: DbSession) -> None:
     """
     Edits the attributes of a workflow.
     """
+
+    if not can_access_workflow(auth, workflow, session, for_writing=True):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
     if params.name is not None:
         workflow.name = params.name
@@ -120,27 +147,29 @@ async def edit(params: WorkflowUpdateParams, workflow: ExistingWorkflow) -> None
 
 @router.post("/{workflow_id}/delete",
              summary="Delete Workflow")
-async def delete(workflow: ExistingWorkflow, session: DbSession) -> None:
+async def delete(workflow: ExistingWorkflow, auth: Authenticated, session: DbSession) -> None:
     """
     Deletes a workflow.
     """
+
+    if not can_access_workflow(auth, workflow, session, for_writing=True):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
     session.delete(workflow)
 
 
 @router.post("/{workflow_id}/share/add",
              summary="Share Workflow")
-async def add_share(write: bool, workflow: ExistingWorkflow, user: ExistingUser, session: DbSession) -> None:
+async def add_share(write: bool, workflow: ExistingWorkflow, user: ExistingUser, auth: Authenticated, session: DbSession) -> None:
     """
     Gives read/write rights to a user, or changes the user's current rights.
     """
 
-    share: WorkflowShare | None = session.exec(
-        select(WorkflowShare).where(WorkflowShare.user_id == user.id)
-    ).one_or_none()[0]
-    print(share)
+    if not can_access_workflow(auth, workflow, session, for_writing=True):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
 
-    
+    share = get_workflow_share(user, workflow, session)
+
     if share is None:
         workflow.shares.append(WorkflowShare(
             user_id=user.id,
@@ -153,14 +182,15 @@ async def add_share(write: bool, workflow: ExistingWorkflow, user: ExistingUser,
 
 @router.post("/{workflow_id}/share/remove",
              summary="Un-Share Workflow")
-async def remove_share(workflow: ExistingWorkflow, user: ExistingUser, session: DbSession) -> None:
+async def remove_share(workflow: ExistingWorkflow, user: ExistingUser, auth: Authenticated, session: DbSession) -> None:
     """
     Revokes the right of a user to read from / write to this workflow.
     """
 
-    share: WorkflowShare | None = session.exec(
-        select(WorkflowShare).where(WorkflowShare.user_id == user.id)
-    ).one_or_none()[0]
+    if not can_access_workflow(auth, workflow, session, for_writing=True):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+    
+    share = get_workflow_share(user, workflow, session)
 
     if share is not None:
         session.delete(share)
